feat: Add a "show more" button; Rename element IDs
By default, a maximum of 12 items are shown initially, the rest are hidden until the user clicks "Show more", after which another 12 items are shown. As long as there are hidden items, the button will be available.

I have made sure to reset the menu so that it stays "collapsed" every time another category is selected or search event occurs. Cannot guarantee 100% bug-free code, though.