fix: Signup and Forget Password modal not working
I had to revert the earlier db6d0e3 commit because it inherently broke all modals when opened from an opened modal (I don't know why).

Of course, this also reverted my previous fix for the language dropdown not opening for the first click. Between the two problems however, obviously not being able to recover passwords and/or creating an account is far more serious.