feat: Add food details pop-up
There's still an issue with mobile screens: the modal is designed with full viewport height in mind (no searchbar on mobile). Realistically speaking this is not the case. As the user swipes up and down, the searchbar will show up/hides itself, making the 100% viewport height thing hard to solve without JavaScript. I'll leave it there until there's time to fix it.