fix: Broken cart buttons after adding item to cart
Because the content of cart-table is dynamically generated (using JS), event handlers on such elements can be broken quite easily by adding and removing said elements.

The solution here is to use event delegation, which is to add event listeners to parent elements (that are statically generated - they are present ever since page load). The event listeners will then take an additional element (this will be used to decide which child element triggered the event).

Since the parent element is always there, it can "adapt" to their child elements, and will always work.